+++ library/src/main/java/company/tap/gosellapi/open/interfaces/PaymentDataSource.java
@@ -117,4 +117,10 @@ public interface PaymentDataSource {
     Merchant     getMerchant();
     @Nullable
      String getPaymentDataType();
+
+    /**
+     *  Defines if user wants all cards or specific card types.
+     */
+    @Nullable
+    String getCardType();
 }